requirements.txt
Flask==1.1.2
Flask-JWT-Extended==3.24.1
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
passlib==1.7.2
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.18
urllib3==1.25.9